feat(example-app-improvements): moved components into own directory
