+++ examples/shop-with-blog/client/src/pages/shop/Category/components/ShowMore.js
