+++ examples/shop-with-blog/client/src/pages/shop/Category/components/index.js
@@ -1,0 +1,2 @@
+export * from './ShowMore';
+export * from './ShowingOutOf';
